added Taka's template for tww-gz
Added framework based off of tp-gz's current iteration of their framework. This involves statically loading assets, something that to my knowledge their framework is being redesigned currently to avoid. This framework is based in C++, and should be operable with TWW's Hacking API which is primarily written in C. This is unchecked however, and some things written in the hacking API may need to be rewritten for configurability with a C++ framework.